Shift fn to OnPolicyAgent
