+++ genrl/agents/deep/ppo1/ppo1.py
@@ -6,7 +6,6 @@ import torch.nn as nn  # noqa
 import torch.optim as opt  # noqa
 
 from genrl.agents import OnPolicyAgent
-from genrl.core import compute_returns_and_advantage
 from genrl.utils import get_env_properties, get_model, safe_mean
 
 
@@ -138,9 +137,7 @@ class PPO1(OnPolicyAgent):
             values (:obj:`torch.Tensor`): Values of states encountered during the rollout
             dones (:obj:`list` of bool): Game over statuses of each environment
         """
-        print(type(dones))
-        compute_returns_and_advantage(
-            self.rollout,
+        self.compute_returns_and_advantage(
             values.detach().cpu().numpy(),
             dones.cpu().numpy(),
             use_gae=True,
