+++ genrl/core/rollout_storage.py
@@ -257,51 +257,3 @@ class RolloutBuffer(BaseBuffer):
             self.returns[batch_inds].flatten(),
         )
         return RolloutBufferSamples(*tuple(map(self.to_torch, data)))
-
-
-def compute_returns_and_advantage(
-    rollout_buffer: Union[RolloutBuffer, BaseBuffer],
-    last_value: torch.Tensor,
-    dones: np.ndarray,
-    use_gae: bool = False,
-) -> None:
-    """
-    Post-processing function: compute the returns (sum of discounted rewards)
-    and advantage (A(s) = R - V(S)).
-    Adapted from Stable-Baselines PPO2.
-    ;param rollout_buffer: (RolloutBuffer, BaseBuffer) An instance of the rollout buffer used in On-policy algorithms
-    :param last_value: (torch.Tensor)
-    :param dones: (np.ndarray)
-    :param use_gae: (bool) Whether to use Generalized Advantage Estimation
-        or normal advantage for advantage computation.
-    """
-    last_value = last_value.flatten()
-
-    if use_gae:
-        gae_lambda = rollout_buffer.gae_lambda
-    else:
-        gae_lambda = 1
-
-    temp_dones = torch.cat(
-        [rollout_buffer.dones, torch.as_tensor(dones).unsqueeze(0)], dim=0
-    )
-    temp_values = torch.cat(
-        [rollout_buffer.values, torch.as_tensor(last_value).unsqueeze(0)], dim=0
-    )
-
-    running_advantage = 0.0
-    for step in reversed(range(rollout_buffer.buffer_size)):
-        next_non_terminal = 1 - temp_dones[step + 1]
-        next_values = temp_values[step + 1]
-        delta = (
-            rollout_buffer.rewards[step]
-            + rollout_buffer.gamma * next_non_terminal * next_values
-            - temp_values[step]
-        )
-        running_advantage = (
-            delta
-            + rollout_buffer.gamma * gae_lambda * next_non_terminal * running_advantage
-        )
-        rollout_buffer.advantages[step] = running_advantage
-
-    rollout_buffer.returns = rollout_buffer.advantages + rollout_buffer.values
